move docker win32 code to docker service
